requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==23.2.0
bitsandbytes==0.43.3
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cramjam==2.10.0
cryptography==43.0.1
curl_cffi==0.10.0
cycler==0.12.1
docker-pycreds==0.4.0
docstring_parser==0.16
fastparquet==2024.11.0
filelock==3.13.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.65.4
hydra-core==1.3.2
idna==3.7
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.3
jsonargparse==4.32.0
kiwisolver==1.4.5
lightning==2.3.3
lightning-utilities==0.11.6
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.0.5
networkx==3.2.1
numpy==1.26.3
olefile==0.47
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
perlin-noise==1.13
pillow==10.2.0
pillow-avif-plugin==1.5.1
platformdirs==4.2.2
protobuf==4.25.3
psutil==5.9.8
pyarrow==19.0.1
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.3.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tbb==2021.11.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tqdm==4.66.4
typeshed_client==2.7.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.0
Werkzeug==3.0.3
yarl==1.9.4